EmailController, ForgetPasswordController added & routes optimized
