+++ app/Http/Controllers/Auth/ForgetPasswordController.php
@@ -2,7 +2,12 @@
 
 namespace App\Http\Controllers\Auth;
 
+use App\Jobs\PasswordResetEmailJob;
+use App\Models\User;
 use Carbon\Carbon;
+use Illuminate\Auth\Events\PasswordReset;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Password;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -17,7 +22,56 @@ class ForgetPasswordController extends Controller
         return view('auth.frontend.forget-password');
     }
 
-    public function submitForgetPasswordForm(Request $request)
+    public function sendForgetPasswordLink(Request $request){
+        $request->validate(['email' => 'required|email']);
+
+        PasswordResetEmailJob::dispatch($request->only('email'));
+        toastr()->success('Password reset link will be sent to your email!', 'Reset Link Sent');
+        return back()->with(['status' => 'Reset Link Sent Success!']);
+    }
+
+    public function verifyForgetPasswordLink($token)
+    {
+        return view('auth.frontend.reset-password', ['token' => $token]);
+    }
+
+    public function updateNewPassword(Request $request){
+        $request->validate([
+            'token' => 'required',
+            'email' => 'required|email',
+            'password' => 'required|same:password_confirmation|min:8',
+            'password_confirmation' => 'required|same:password'
+        ]);
+        $status = Password::reset(
+            $request->only('email', 'password', 'password_confirmation', 'token'),
+            function (User $user, string $password) {
+                $user->forceFill([
+                    'password' => Hash::make($password)
+                ])->setRememberToken(Str::random(60));
+                $user->save();
+                event(new PasswordReset($user));
+            }
+        );
+
+        if ($status == Password::PASSWORD_RESET) {
+            toastr()->success('Password has been changed!', 'Successfully Changed');
+            return redirect()->route('user_login')->with('status', __($status));
+        } else {
+            toastr()->error('Failed to change password!', 'Failed Change');
+            return back()->withErrors(['email' => [__($status)]]);
+        }
+    }
+
+
+    /*
+     * Send reset password link to the given user email.
+     *
+     * (
+     *  Currently not in used.
+     * Only to show demo for our presentation
+     *  )
+     */
+    public function mailForgetPasswordLink(Request $request)
     {
         $request->validate([
             'email' => 'required|email|exists:users',
@@ -38,4 +92,5 @@ class ForgetPasswordController extends Controller
 
         return back()->with('message', 'We have e-mailed your password reset link!');
     }
+
 }
